+++ test/test.api.js
@@ -275,6 +275,38 @@ describe('make sure the server is running', function() {
         });
     });
 
+    describe('/api/calendar.json?types=cargotram is working', function() {
+        it('should return something', function(done) {
+            supertest
+                .get('/api/calendar.json?types=cargotram')
+                .expect(200, done);
+        });
+    });
+
+    describe('/api/calendar.json?types[]=cargotram is working', function() {
+        it('should return something', function(done) {
+            supertest
+                .get('/api/calendar.json?types[]=cargotram')
+                .expect(200, done);
+        });
+    });
+
+    describe('/api/calendar.json?types[]=cargotram&types[]=etram is working', function() {
+        it('should return something', function(done) {
+            supertest
+                .get('/api/calendar.json?types[]=cargotram&types[]=etram')
+                .expect(200, done);
+        });
+    });
+
+    describe('/api/calendar.json?types=cargotram&types=etram is working', function() {
+        it('should return something', function(done) {
+            supertest
+                .get('/api/calendar.json?types=cargotram&types=etram')
+                .expect(200, done);
+        });
+    });
+
     // kill the server again
     after(function() {
         server.stop();